Update RAK811.cpp
Add a finishing zero to the compositionBuffer

The compositionBuffer needs a last zero to indicate the end of the string to send.

Right now the string that is sent ends with a random number of extra characters. This makes it impossible for Cayenne to decode the payload and might perhaps crash the device.